PBRF.jl

PBRF.jl is a Julia package for calculating polarized bidrectional reflectance factors. It is largely based on M. Mishchenko et al.'s code.

PBRF.Wrapper.run_pbrf

PBRF.evaluate

PBRF.get_itp

PBRF.evaluate — Method

Evaluate the refletion matrix $\mathbf{R}(\mu,\phi;\mu_0,\phi_0)$ using built interpolations.

evaluate(itps, μ, ϕ, μ₀, ϕ₀)

PBRF.get_itp — Method

Build interpolations from the calculated Fourier coefficients.

get_itp(x, R)

PBRF.Wrapper.run_pbrf — Method

Calculate the Fourier coefficients of the refletion matrix using PBRF.

run_pbrf(ω, ngauss, coeff; epsilon, mode)

ω is the single scattering albedo, which should be within (0, 1].

ngauss determines the number of points used in integration.

coeff is the expansion coefficients, which should be an R x 6 matrix.

Optional parameters:

epsilon is the threshold for convergence check. Detault is 1e-7.

mode determines the way to do the quadrature. Default is PBRF.Standard (NQUADR=2 in the original code), which is a normal Gaussian-Legendre quadrature within [0, 1]. Other options are PBRF.NormalOriented (NQUADR=1 in the original code) and PBRF.NeedNormal (NQUADR=3 in the original code). Note that PBRF.Standard might not be suitable if the refletion in the normal direction is needed.

Results:

x is the quadrature nodes.

R is a 4 x 4 x NG x NG x LMAX1 array, storing all the Fourier coefficients.